refactor(batch): add and use svelte component for results
improve functions that are used on the batch page and make them more efficient
