--- src/lib/schema/batch.ts
@@ -1,8 +1,0 @@
-export interface SourceData {
-    campaign: string;
-}
-
-
-interface SidebarRowData {
-    campaign: string;
-}
